gptel-curl: Only convert to Org in Org buffers
* gptel-curl.el (gptel-curl-get-response): Don't convert response
into org-mode unless the buffer from which the request originated
is in org-mode.  This makes `gptel-default-mode' less binding, and
only used when creating a new chat session with `gptel'.  Also,
gptel should now do the right thing depending on whether the
current buffer is in text, Markdown or Org modes.