Add support for mutual TLS
I would like to use GPTel with an OpenAI gateway that requires mutual TLS.

This change adds 2 new slots to the `gptel-backend` struct to support
mutual TLS:

- The optional `client-certificate` slot is used to set the filename
of the TLS client certificate.

- The optional `client-private-key` slot is used to set the filename
of the private key of the TLS client.

The client certificate and private key are then used in the curl
backend via the `--cert` and `--key` options when making HTTP
requests.