Add curl-args slot to gptel-backend
This slot is used to append backend specific options to the curl
arguments used to make HTTP requests. The backend specific arguments
are appended after `gptel-curl--common-args`.

My use case for this feature is to set the `--cert` and `--key`
options in a custom backend that uses mutal TLS to communicate with an
OpenAI proxy/gateway.