gptel asks user to select a buffer if many are open
This commit modifies the behavior of the gptel function so that, in
the case where the function is called interactively, and multiple
gptel sessions already exist (as judged by there being multiple
buffers with the minor mode gptel-mode enabled), then the function
prompts the user to select one of those buffers.

This replaces the previous behavior in this case, which was just to
switch to the buffer named `backend-name` if it already existed, or
else to create a new buffer with that name.

So, after this comit, the gptel function behaves as follows:

- (NEW). When called interactively and multiple gptel-mode buffers exists, it
  prompts the user to choose a buffer to switch to.

- when called interactively and no gptel-mode buffers exist, it
  creates a buffer named `backend-name`

- when called interactively and 1 gptel-mode buffers exists, it
  switches to the buffer named `backend-name` or creates it if it does
  not exist.

- when called interactively with a prefix arg, it behaves as before,
  and prompts the user to acknowledge or input the name of a buffer to create.

- when called programmatically, it behaves as before.