gptel: Include backend in process info plist
* gptel-curl.el (gptel-curl-get-response): Include the relevant
`gptel-backend` in the INFO plist used by the output parsers.
This is so they can access backend-specific settings.

The backend is also passed directly to the response parsers and
used for dispatch, to find the right defmethod, but this is
unreliable and should not be used.  Specifically, passing this is
omitted when using Curl for stream processing, since we do the
dispatch early, when sending the request.

* gptel.el (gptel--url-parse-response): Ditto, for when using the
`url-retrieve` method.