gptel-rewrite: New user option for rewrite message
* gptel-rewrite.el (gptel-rewrite-directives-hook,
gptel-rewrite-menu): New user option
`gptel-rewrite-directives-hook` for customizing the default
rewrite/refactor directive.  Each function in the hook is called
until one returns a non-nil value, the string to use as the
directive in the current context.  This can be used to customize
the refactoring prompt based on any predicate, such as the major
mode or the contents of the buffer. (#370)