gptel-context: Include images via `gptel-add'
* gptel-context.el (gptel-context--buffer-setup,
gptel-context--string, gptel-context--collect,
gptel-context--collect-media, gptel-context--wrap,
gptel-context--make-overlay, gptel-context-remove,
gptel--file-binary-p, gptel-context-add,
gptel-context-wrap-function): Supported media files (currently
only images) can now be added to the context with
`gptel-add-file', or with `gptel-add' in an image-mode buffer.
When previewing the current context in the context buffer, preview
images correctly.  New functions `gptel-context--collect-media' to
collect all media files in the context for processing, and
`gptel--file-binary-p' to check if a file is binary.

* gptel.el (gptel--create-prompt): When available, add media files
to context when creating the prompt.

* gptel-openai.el (gptel--openai-parse-multipart,
gptel--wrap-user-prompt): Wrap the text and media parts of the
context separately into the system message or user prompt.  (Media
files are always added to the user prompt, as some backends do not
support them in the system message.)

* gptel-ollama.el (gptel--ollama-parse-multipart,
gptel--wrap-user-prompt): Ditto.

* gptel-anthropic.el (gptel--anthropic-parse-multipart,
gptel--wrap-user-prompt): Ditto.