gptel-anthropic: Ignore blank user prompts
* gptel-anthropic.el (gptel--parse-buffer): Ignore blank user
promps in the buffer when creating the full prompt.  The most
common way to unintentionally produce these blank regions is to
reformat the response in some way, such as by filling the response
region.  Until a better solution for tracking responses can be
found, this will have to do.  Address #409, #406 and #351.  Also
relevant to the discussion about overlay-based tracking in #321.