gptel-curl: don't try converting CR-LF on Windows
We use curl's header size output to `goto-char` to the end of the response
header (to parse the JSON that starts there).

On Windows, however, processes we launch will typically output Windows-style
line endings, with CR-LF; by default, Emacs tries to convert this to just CR,
to avoid seeing ugly `^M` characters at the end of lines.

 `curl` dumps what it gets from the server directly... but HTTP also specifies
 CR-LF line endings, at least at the end of headers. On Windows, Emacs converts
 these to CRs, removing the LFs, and thus causing our header size measurement to
 be slightly off (the buffer now includes fewer characters of header vs. what
 `curl` told us we'd have). So `goto-char` seeks too much ahead, missing the
 beginning of the JSON; we get a JSON parse error.

This PR fixes this by forcing the process coding system to the UNIX
variant (with no CR-LF conversion).

We gate this behind an "is this Windows" check so it shouldn't affect other
systems.

Fixes https://github.com/karthink/gptel/issues/455.