gptel: Support model APIs without system messages
Add support for models that don't use system messages.  By default
gptel assumes that chat models use system messages (also called
system prompts or system instructions).  To indicate that a model
does not, we add the capability `nosystem' to the model's
capabilities list.  This specification is "inverted" since most
models support system messages, and we would like to avoid having
to specify this every time.  This method of configuration is still
experimental and subject to change.

Add support for unsetting a system message, i.e. setting it to
nil.  This makes it possible to use models with their default
system prompts that are baked into their API.  Ensure that gptel's
transient menu correctly displays and distinguishes between (i) an
unset system message and (ii) lack of system message support for
the active model.

In preparation for more model-specific options in gptel's
transient menus, we refresh the transient menu automatically after
changing the model.  This is a full refresh since Transient does
not support more granular reactivity.

* gptel.el (gptel--create-prompt, gptel-request): Handle models
that don't support system messages correctly, in `gptel-request'
and with regards to adding context chunks.  If `gptel-use-context'
is set to `system' but the model does not support system messages,
add the context to the user prompt instead.

* gptel-transient.el (gptel-system-prompt--setup,
gptel--format-system-message, gptel--get-directive,
transient-infix-set, gptel-system-prompt--setup,
gptel-system-prompt, gptel--suffix-send,
gptel--suffix-system-message): Handle the display of "no system
message" and "no support for system message" in the transient
menus.  New helper function `gptel--format-system-message' to
format the system message (or equivalent) for display.  Adjust the
handling of the "additional directive" option so that it works
correctly when the system message is unset.  The "additional
directive" becomes the totality of the system message in this
case.  For models that don't support system messages, the
"additional directive" option no longer appears.

* gptel-openai.el (gptel--process-models, gptel--parse-buffer):
When creating model symbol-plists from model specifications,
overwrite the existing symbol-plist instead of merging.  Without
this, no keys can be removed from the plist, making it difficult
to reset model capability descriptions etc.  Fix system message
support per above description.

* gptel-ollama.el (gptel--parse-buffer): Fix system message
support per above description.

* gptel-gemini.el (gptel--request-data, gptel--parse-buffer): Fix
system message support per above description.  Even though the
model `gemini-pro' does not support system messages, we continue
to "fake" this by prepending the user prompt with the system
message to minimize the disruption to gemini Users.  This behavior
might be changed in the future.

* gptel-anthropic.el (gptel--request-data): Fix system message
support per above description.