gptel: Add generic gptel--parse-list function
* gptel.el:
(gptel--parse-list): Generic function to parse a list of strings
into a conversation prompt, assigning alternating user and
assistant roles to the strings.  The form of the prompt is
API-specific and implementations for each backend type are needed.
This is useful to construct and send "synthetic"
prompts (templated conversations) using `gptel-request'.

* gptel-openai.el (gptel--parse-list): Add implementation for
OpenAI.

* gptel-ollama.el (gptel--parse-list): Add implementation for
Ollama.

* gptel-gemini.el (gptel--parse-list): Add implementation for
Gemini.

* gptel-anthropic.el (gptel--parse-list): Add implementation for
Anthropic.