gptel: Generalize gptel directives
* gptel.el (gptel-directives, gptel-request,
gptel--strip-mode-suffix, gptel--describe-directive,
gptel--parse-directive, gptel-mode): Generalize the notion of
gptel's directives to include dynamic system messages and
templates.  Now entries in `gptel-directives' can be
- strings, as before: these are interpreted as system messages (no
change)
- lists of strings: the head is the system message, and the
tail is an alternating series of user/llm role messages.
This can serve as a static template of the intial part of the
conversation.
- functions: The function must return a string or list of strings,
which are interpreted as the above.  This can be used to construct
system messages and conversations dynamically.

`gptel-request' can now optionally take
- a list of strings as its :prompt argument, corresponding to
alternating user/llm role messages.
- a general gptel directive as its :system argument (i.e. a
string, list of strings or a function)

* gptel-transient.el (gptel--merge-additional-directive,
gptel-system-prompt--setup, gptel--format-system-message,
gptel-system-prompt--format, gptel-menu, gptel-system-prompt,
gptel--suffix-send, gptel--suffix-system-message): Adjust
transient features to display the new types of directives
correctly.  Interactively editing directives that are functions
and lists of strings is not supported: only the system message
part (of lists) or the results of calling the directive (when it
is a function) can be edited.