gptel: Adjust persistence for dynamic directives
* gptel.el (gptel--save-state): When saving state, save the static
value of the active gptel directive.  This means dynamic
directives will be evaluated, and only the system message part of
directives that are templated conversations will be saved to the
file.

It's not clear yet how to correctly persist dynamic directives.
This limitation may be addressed in the future.

* gptel-org.el (gptel-org-set-properties): Similar adjustments for
Org mode chat buffers.