gptel-rewrite: Rewrite directive handling changes
gptel's rewrite directive handling has been inconsistent with the
rest of the package, and poor as well.  Fix by making rewrite
directives a first-class concept, on par with regular
directives (i.e. system messages or templates.)

Improve the UI of the rewrite menu.

NOTE: For consistency, many of these function names will change in
the following commits.

* gptel-transient.el:
(gptel--read-with-prefix-help, gptel--read-with-prefix): Function
to read from the minibuffer while showing a prefix string.  Used
to show the full directive (general or when rewriting) to help
write additional instructions.
(gptel-system-prompt--setup, gptel--setup-directive-menu):
Function that generates the transient infixes required for
selecting a directive.  This used to be part of
`gptel-system-prompt--setup' but has been generalized for reuse.
(gptel-system-prompt):  Changes for compatibility.
(gptel--infix-add-directive):  Use the read-with-prefix helper
function.
(gptel--suffix-system-message, gptel--edit-directive): Extract the
code for editing a string (the directive) in a separate buffer
into `gptel--edit-directive'.

* gptel-rewrite.el:
(gptel--rewrite-expand-prompt): This function, used to preview a
string when reading from the minibuffer, is more generally useful.
Move to gptel-transient.el as `gptel--read-with-prefix'.
(gptel-rewrite-directive): New variable serving the same function
as `gptel--system-message'.  It is a directive (i.e. system
message or template) used when rewriting/refactoring.  Its value
can be set by choosing amongst `gptel-directives', and
user-defined ones can be added.  Its default value corresponds to
the key 'rewrite' in `gptel-directives'.
(gptel--rewrite-directive-default): Default value of
gptel--rewrite-directive.  Renamed from
`gptel--rewrite-directive'.
(gptel-rewrite-menu, gptel--rewrite-directive-menu): UI changes.
Add a menu to select the rewrite directive.  Add the infix
`gptel--infix-context-remove-all' to remove all context
(gptel--infix-rewrite-prompt, gptel--infix-rewrite-extra): Rename.
(gptel--rewrite-infix-diff:-U, gptel--infix-rewrite-diff:-U):
Rename for consistency.
(gptel--suffix-rewrite-directive): Suffix to edit the rewrite directive.
(gptel--suffix-rewrite): Use `gptel--rewrite-directive' and
`gptel--rewrite-message' to create the full rewrite prompt.