gptel-curl: Changes to the streaming API
* gptel-curl.el (gptel-curl--stream-insert-response,
gptel-curl--stream-cleanup, gptel-curl-get-response): To make
`gptel-request' easier to use with streaming responses, the
following changes have been made:

1. The INFO plist passed to gptel-request's callback now contains
a :stream key that is set if the current request is a streaming
request:
(plist-get info :stream) ;=> t

2. When using a callback with a streaming request,

(lambda (response info) ...),

gptel will call it with a response value t for a successful
finish, and with a response value of `error' for a failed request.
Callers can dispatch on the type of response to handle cleanup for
streaming requests.  NOTE: This API is still experimental.

There are no changes required to `url-retrieve' based requests,
since they don't support streaming.