gptel: Add finite state machine
* gptel.el (gptel-fsm, gptel--fsm-transition, gptel--fsm-next,
gptel-request--transitions, gptel-request--handlers): Add a state
machine and state transition system to gptel.  While the
implementation is generic, this will be used
to drive (and be driven by) `gptel-request'.  This allows us to
encode more dynamic control flows, such as multi-turn requests
involving tool-use, without creating a mess inside the process
filters and sentinel logic.  This is not hooked up to
`gptel-request' yet.

For now gptel uses `fsm' to denote this Finite State Machine.
This API is experimental and its implementation may change in the
future.

Add an (empty) default state transition table and state handler
list for `gptel-fsm'.  The transition table encodes possible state
transitions along with predicate functions that determine the next
state.  The handler list contains functions that are called when
entering each state.  These will be populated next.