gptel: Add tool call parsing (only) to most backends
* gptel-openai.el (gptel--json-read-string,
gptel-curl--parse-stream, gptel--parse-response):
New macro `gptel--json-read-string' to read a json string.

Add tool call parsing for both streaming and one-shot responses.
The new functions are considerably longer, but the "happy path"
for text is the same.  The additional complexity is all towards
collating tool calls, which are recorded in the state machine's
INFO.

* gptel-gemini.el (gptel-curl--parse-stream,
gptel--parse-response): Ditto.

* gptel-anthropic.el (gptel-curl--parse-stream,
gptel--parse-response): Ditto.

* gptel-ollama.el (gptel--parse-response): Ditto, but only for
non-streaming responses.  Ollama does not yet support tool calls
with streaming responses.

* gptel-privategpt.el (gptel--privategpt-parse-stream): Tool call
parsing is not implemented: add note for the future.

Note: tool-call processing is not yet added to gptel, so the
captured tool-calls do nothing.