Skip reformatting markdown to org in src-blocks
As is, reformatting code will replace " ** "'s with "*", and it will replace
comments with headers.

This change changes the temp buffer to org-mode, and skips replacements
if they are inside of source blocks.

This works because the backticks are replaced with source blocks before
all of the other formatting.

The issue here can be tested by:
- asking gptel to return the python function lambda x: x **
2.  It will return x * 2. The spacing is important, x**2 will not get converted.
- asking gptel to return some python code with comments, e.g.
\# this will become an org mode bullet
print(1)