gptel: Change JSON parsing semantics
* gptel-openai.el (gptel--json-encode, gptel--json-read):
Distinguish between false, null and the empty JSON object in Elisp
by reading them as :json-false, :null and nil (which is the same
as '()).  Formerly, we were conflating null and the empty JSON
object.  This distinction is required for handling tool call
arguments correctly.

As of this commit, all the parsing functions (except Kagi) are
broken as a result.  This will be addressed in the next few commits.

* gptel-kagi.el (gptel--parse-response): Handle :null output.