gptel: Add gptel-post-request-hook
* gptel.el (gptel-request, gptel-post-request-hook): Add hook that
runs after the request is sent, and (possibly) before any response
is received.  The intent of this hook is to run preparatory or
reset code.  One application of this is the (forthcoming)
implementation of "oneshot" settings, where model parameters can
be set from the transient menu for the next response only.