gptel-transient: Add "oneshot" scope to set vars
* gptel-transient.el (gptel--set-with-scope, gptel--scope,
transient-infix-read, transient-format-value,
gptel--infix-variable-scope): Add a "oneshot" scope for setting
model parameters and the system message.  By flipping this switch
in gptel's transient menu, you can set these variables for the
next request only, after which they reset to their current values.
This is controlled by the (still internal) variable
`gptel--set-buffer-locally', which now controls whether relevant
parameters are set globally, buffer-locally or in a oneshot
manner.

The name of this variable is intended to be changed in the future.