gptel: Simplify gptel--request-data
We are now handling the determination of request
parameters (whether we should use streaming, and a system message)
in `gptel-request'.  This means we do not need to check for these
things in the various implementations of `gptel--request-data'.

* gptel-privategpt.el (gptel--request-data): Simplify handling of
streaming, system message.

* gptel-openai.el (gptel--request-data): Simplify handling of
streaming, system message.

* gptel-ollama.el (gptel--request-data): Simplify handling of
streaming, system message.

* gptel-gemini.el (gptel--request-data): Simplify handling of
streaming, system message.

* gptel-anthropic.el (gptel--request-data): Simplify handling of
streaming, system message.